Added date selection capabilities
